+++ variants/models.py
@@ -1256,6 +1256,10 @@ class SmallVariantFlags(models.Model):
     flag_visual = models.CharField(
         max_length=32, choices=VARIANT_RATING_CHOICES, default="empty", null=False
     )
+    #: Molecular flag.
+    flag_molecular = models.CharField(
+        max_length=32, choices=VARIANT_RATING_CHOICES, default="empty", null=False
+    )
     #: Wet-lab validation flag.
     flag_validation = models.CharField(
         max_length=32, choices=VARIANT_RATING_CHOICES, default="empty", null=False
@@ -1299,7 +1303,7 @@ class SmallVariantFlags(models.Model):
             for name in ("bookmarked", "for_validation", "candidate", "final causative"):
                 if getattr(self, "flag_%s" % name.replace(" ", "_")):
                     flag_desc.append(name)
-            for name in ("visual", "validation", "phenotype_match", "summary"):
+            for name in ("visual", "validation", "molecular", "phenotype_match", "summary"):
                 field = getattr(self, "flag_%s" % name)
                 if field and field != "empty":
                     flag_desc.append("%s rating is %s" % (name.split("_")[0], field))
@@ -1322,6 +1326,7 @@ class SmallVariantFlags(models.Model):
                 self.flag_candidate,
                 self.flag_final_causative,
                 self.flag_for_validation,
+                self.flag_molecular != "empty",
                 self.flag_visual != "empty",
                 self.flag_validation != "empty",
                 self.flag_phenotype_match != "empty",